Clean up unused workflows
We continuously deploy to https://compound.element.io using Cloudflare Pages, so we don't need to deploy to GitHub Pages as well. However, the workflow that did that was still useful for verifying that this actually builds, so I've kept that part.

Also, we don't release anything here to NPM.